+++ Automate/Framework/MachineGroup.cs
@@ -1,13 +1,26 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using Microsoft.Xna.Framework;
+using Pathoschild.Stardew.Common.Utilities;
 using StardewValley;
+using SObject = StardewValley.Object;
 
 namespace Pathoschild.Stardew.Automate.Framework
 {
     /// <summary>A collection of machines and storage which work as one unit.</summary>
     internal class MachineGroup
     {
+        /*********
+        ** Fields
+        *********/
+        /// <summary>The number of milliseconds to pause machines when they crash.</summary>
+        private readonly int ErrorPauseMilliseconds = 30000;
+
+        /// <summary>Machines which are temporarily paused, with the game time in milliseconds when their pause expires.</summary>
+        private readonly IDictionary<IMachine, double> MachinePauseExpiries = new Dictionary<IMachine, double>(new ObjectReferenceComparer<IMachine>());
+
+
         /*********
         ** Accessors
         *********/
@@ -50,11 +63,26 @@ namespace Pathoschild.Stardew.Automate.Framework
         /// <summary>Automate the machines inside the group.</summary>
         public void Automate()
         {
+            double curTime = Game1.currentGameTime.TotalGameTime.TotalMilliseconds;
+
+            // clear expired timers
+            if (this.MachinePauseExpiries.Count > 0)
+            {
+                foreach (var entry in this.MachinePauseExpiries.ToArray())
+                {
+                    if (curTime >= entry.Value)
+                        this.MachinePauseExpiries.Remove(entry.Key);
+                }
+            }
+
             // get machines ready for input/output
             IList<IMachine> outputReady = new List<IMachine>();
             IList<IMachine> inputReady = new List<IMachine>();
             foreach (IMachine machine in this.Machines)
             {
+                if (this.MachinePauseExpiries.ContainsKey(machine))
+                    continue;
+
                 switch (machine.GetState())
                 {
                     case MachineState.Done:
@@ -72,8 +100,30 @@ namespace Pathoschild.Stardew.Automate.Framework
             // process output
             foreach (IMachine machine in outputReady)
             {
-                if (this.StorageManager.TryPush(machine.GetOutput()) && machine.GetState() == MachineState.Empty)
-                    inputReady.Add(machine);
+                ITrackedStack output = null;
+                try
+                {
+                    output = machine.GetOutput();
+                    if (this.StorageManager.TryPush(output) && machine.GetState() == MachineState.Empty)
+                        inputReady.Add(machine);
+                }
+                catch (Exception ex)
+                {
+                    string error = $"Failed to automate machine '{machine.MachineTypeID}' at {machine.Location?.Name} (tile: {machine.TileArea.X}, {machine.TileArea.Y}). An error occurred while ";
+                    if (output == null)
+                        error += "retrieving its output.";
+                    else
+                    {
+                        error += $"storing its output item #{output.Sample.ParentSheetIndex} ('{output.Sample.Name}'";
+                        if (output.Sample is SObject outputObj && outputObj.preservedParentSheetIndex.Value >= 0)
+                            error += $", preserved item #{outputObj.preservedParentSheetIndex.Value}";
+                        error += ").";
+                    }
+                    error += $" Machine paused for {this.ErrorPauseMilliseconds / 1000}s.";
+
+                    this.MachinePauseExpiries[machine] = curTime + this.ErrorPauseMilliseconds;
+                    throw new InvalidOperationException(error, ex);
+                }
             }
 
             // process input